FIX: dockerfile

+++ build.sh
@@ -13,7 +13,13 @@ then
     cd frontend
 fi
 
-rm -rf build
+if [[ -d /build ]]
+then
+    echo "/build exists on your filesystem."
+    rm -rf build
+fi
+
+
 
 ## Run rest of the commands if we're already in the frontend directory
